+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHATC7R18YPDDQJ5AC9PDK8H"
+        "01KHB0WMCAWJG8NQ8HEV6VB8PV"
     )
     r"""The unique identifier of the data source"""