+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHATC7QZ0P3P6SZD2GM9PPZS"
+        "01KHB0WMBYNNYPP5RQ26HWDV16"
     )
     r"""The unique identifier of the data source"""