+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -2093,7 +2093,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KHATC71B4G87WMW6X6CNAA4Y"
+    id: Optional[str] = "01KHB0WKQ4XPXF7ENHGVEDE7ZF"
 
     description: Optional[str] = None